+++ selfdrive/controls/lib/lateral_planner.py
@@ -140,7 +140,7 @@ class LateralPlanner():
       self.auto_lane_user_enabled = False
     
     md = sm['modelV2']
-    activate_auto_lane_pos = self.LP.lane_offset.do_auto_enable(int(sm['liveMapData'].currentRoadType)) if self.auto_auto_lane_pos_enabled else AUTO_AUTO_LANE_MODE.NO_CHANGE
+    activate_auto_lane_pos = self.LP.lane_offset.do_auto_enable() if self.auto_auto_lane_pos_enabled else AUTO_AUTO_LANE_MODE.NO_CHANGE
     if activate_auto_lane_pos == AUTO_AUTO_LANE_MODE.ENGAGE and not self.auto_lane_user_disabled:
       put_nonblocking("AutoLanePositionActive", "1")
       self.auto_lane_pos_active = True
@@ -412,6 +412,11 @@ class LateralPlanner():
     plan_send.lateralPlan.roadEdgeProbs = [float(i) for i in self.LP.lane_offset._road_edge_probs]
     plan_send.lateralPlan.trafficLeft = LANE_TRAFFIC.to_cereal(self.LP.lane_offset._left_traffic)
     plan_send.lateralPlan.trafficRight = LANE_TRAFFIC.to_cereal(self.LP.lane_offset._right_traffic)
+    plan_send.lateralPlan.trafficCountLeft = self.LP.lane_offset._left_traffic_count
+    plan_send.lateralPlan.trafficCountRight = self.LP.lane_offset._right_traffic_count
+    plan_send.lateralPlan.trafficMinSeperationLeft = float(self.LP.lane_offset._left_traffic_min_sep_dist.x)
+    plan_send.lateralPlan.trafficMinSeperationRight = float(self.LP.lane_offset._right_traffic_min_sep_dist.x)
+    plan_send.lateralPlan.laneDistFromCenter = float(self.LP.lane_dist_from_center.x)
     if self.auto_lane_pos_active:
       if self.LP.lane_offset._lane_pos_auto == -1. and self.lane_pos != -1.:
         self.lane_pos = -1.