+++ selfdrive/controls/lib/vision_turn_controller.py
@@ -2,9 +2,10 @@ import numpy as np
 import math
 from cereal import log
 from collections import defaultdict
+from common.filter_simple import FirstOrderFilter
 from common.numpy_fast import interp
 from common.params import Params
-from common.realtime import sec_since_boot
+from common.realtime import sec_since_boot, DT_MDL
 from selfdrive.config import Conversions as CV
 from selfdrive.controls.lib.lane_planner import TRAJECTORY_SIZE
 from selfdrive.controls.lib.drive_helpers import V_CRUISE_MAX
@@ -27,17 +28,17 @@ _EVAL_START = 20.  # mts. Distance ahead where to start evaluating vision curvat
 _EVAL_LENGHT = 150.  # mts. Distance ahead where to stop evaluating vision curvature.
 _EVAL_RANGE = np.arange(_EVAL_START, _EVAL_LENGHT, _EVAL_STEP)
 
-_A_LAT_REG_MAX = 1.55  # Maximum lateral acceleration
+_A_LAT_REG_MAX = 1.65  # Maximum lateral acceleration
 
 # Lookup table for the minimum smooth deceleration during the ENTERING state
 # depending on the actual maximum absolute lateral acceleration predicted on the turn ahead.
-_ENTERING_SMOOTH_DECEL_V = [0.0, -0.2, -2.0]  # min decel value allowed on ENTERING state
-_ENTERING_SMOOTH_DECEL_BP = [1.2, 1.35, 2.3]  # absolute value of lat acc ahead
+_ENTERING_SMOOTH_DECEL_V = [0.0, -0.2, -1.5]  # min decel value allowed on ENTERING state
+_ENTERING_SMOOTH_DECEL_BP = [1.2, 1.425, 2.5]  # absolute value of lat acc ahead
 
 # Lookup table for the acceleration for the TURNING state
 # depending on the current lateral acceleration of the vehicle.
 _TURNING_ACC_V = [0.6, 0.0, -1.5]  # acc value
-_TURNING_ACC_BP = [0.8, 1.5, 2.0]  # absolute value of current lat acc
+_TURNING_ACC_BP = [0.8, 1.6, 2.2]  # absolute value of current lat acc
 
 _LEAVING_ACC = 0.7  # Confortable acceleration to regain speed while leaving a turn.
 
@@ -112,7 +113,7 @@ class VisionTurnController():
     self._v_cruise_setpoint = 0.
     self._v_ego = 0.
     self._a_ego = 0.
-    self._a_target = 0.
+    self._a_target = FirstOrderFilter(0., 1.0, DT_MDL)
     self._v_overshoot = 0.
     self._state = VisionTurnControllerState.disabled
     self._CS = None
@@ -136,7 +137,7 @@ class VisionTurnController():
 
   @property
   def a_target(self):
-    return self._a_target if self.is_active else self._a_ego
+    return self._a_target.x if self.is_active else self._a_ego
 
   @property
   def v_turn(self):
@@ -420,7 +421,7 @@ class VisionTurnController():
       a_target = _LEAVING_ACC
 
     # update solution values.
-    self._a_target = a_target
+    self._a_target.update(a_target)
 
   def update(self, enabled, v_ego, a_ego, v_cruise_setpoint, sm):
     self._op_enabled = enabled