+++ selfdrive/ui/paint.cc
@@ -1021,6 +1021,63 @@ static void ui_draw_measures(UIState *s){
             snprintf(unit, sizeof(unit), "%sF", deg);
             snprintf(name, sizeof(name), "AMB TEMP");}
             break;
+          
+          case UIMeasure::INTERACTION_TIMER: 
+            {
+            int s = scene.controls_state.getInteractionTimer();
+            if (s < 5){
+              val_color = nvgRGBA(255, 125, 100, 200);
+            }
+            int h = s / 3600;
+            s = s % 3600;
+            int m = s / 60;
+            s = s % 60;
+            if (h > 0){
+              snprintf(val, sizeof(val), "%d:%02d:%02d", h, m, s);
+            }
+            else{
+              snprintf(val, sizeof(val), "%d:%02d", m, s);
+            }
+            snprintf(name, sizeof(name), "INTERACT");}
+            break;
+
+          case UIMeasure::INTERVENTION_TIMER: 
+            {
+            int s = scene.controls_state.getInterventionTimer();
+            if (s < 5){
+              val_color = nvgRGBA(255, 125, 100, 200);
+            }
+            int h = s / 3600;
+            s = s % 3600;
+            int m = s / 60;
+            s = s % 60;
+            if (h > 0){
+              snprintf(val, sizeof(val), "%d:%02d:%02d", h, m, s);
+            }
+            else{
+              snprintf(val, sizeof(val), "%d:%02d", m, s);
+            }
+            snprintf(name, sizeof(name), "INTERVENE");}
+            break;
+          
+          case UIMeasure::DISTRACTION_TIMER: 
+            {
+            int s = scene.controls_state.getDistractionTimer();
+            if (s < 5){
+              val_color = nvgRGBA(255, 125, 100, 200);
+            }
+            int h = s / 3600;
+            s = s % 3600;
+            int m = s / 60;
+            s = s % 60;
+            if (h > 0){
+              snprintf(val, sizeof(val), "%d:%02d:%02d", h, m, s);
+            }
+            else{
+              snprintf(val, sizeof(val), "%d:%02d", m, s);
+            }
+            snprintf(name, sizeof(name), "DISTRACT");}
+            break;
             
           case UIMeasure::CPU_TEMP_AND_PERCENTC: 
             {
@@ -1173,7 +1230,7 @@ static void ui_draw_measures(UIState *s){
               scene.altitudeUblox = data2.getAltitude();
               scene.gpsAccuracyUblox = data2.getAccuracy();
             }
-            snprintf(name, sizeof(name), "ALTITUDE");
+            snprintf(name, sizeof(name), "ELEVATION");
             if (scene.gpsAccuracyUblox != 0.00) {
               float tmp_val;
               if (s->is_metric) {
@@ -1577,7 +1634,90 @@ static void ui_draw_measures(UIState *s){
             snprintf(val, sizeof(val), "%.1f", sm["longitudinalPlan"].getLongitudinalPlan().getVisionMaxPredictedLateralAcceleration());
             snprintf(unit, sizeof(unit), "m/s²");
             break;}
+
+          case UIMeasure::LANE_POSITION:
+            {
+            snprintf(name, sizeof(name), "LANE POS");
+            auto dat = scene.lateral_plan.getLanePosition();
+            if (dat == LanePosition::LEFT){
+              snprintf(val, sizeof(val), "left");
+            }
+            else if (dat == LanePosition::RIGHT){
+              snprintf(val, sizeof(val), "right");
+            }
+            else{
+              snprintf(val, sizeof(val), "center");
+            }
+            break;}
+
+          case UIMeasure::LANE_OFFSET:
+            {
+            snprintf(name, sizeof(name), "LN OFFSET");
+            auto dat = scene.lateral_plan.getLaneOffset();
+            snprintf(val, sizeof(val), "%.1f", dat);
+            snprintf(unit, sizeof(unit), "m");
+            break;}
           
+          case UIMeasure::TRAFFIC_COUNT_TOTAL:
+            {
+            snprintf(name, sizeof(name), "TOTAL");
+            int dat = scene.lead_vertices_oncoming.size()
+                      + scene.lead_vertices_ongoing.size()
+                      + scene.lead_vertices_stopped.size()
+                      + (scene.lead_data[0].getStatus() ? 1 : 0)
+                      + (scene.radarState.getLeadsCenter()).size();
+            snprintf(val, sizeof(val), "%d", dat);
+            snprintf(unit, sizeof(unit), "cars");
+            break;}
+
+          case UIMeasure::TRAFFIC_COUNT_ONCOMING:
+            {
+            snprintf(name, sizeof(name), "ONCOMING");
+            int dat = scene.lead_vertices_oncoming.size();
+            snprintf(val, sizeof(val), "%d", dat);
+            snprintf(unit, sizeof(unit), "cars");
+            break;}
+
+          case UIMeasure::TRAFFIC_COUNT_ONGOING:
+            {
+            snprintf(name, sizeof(name), "ONGOING");
+            int dat = scene.lead_vertices_ongoing.size()
+                      + (scene.lead_data[0].getStatus() ? 1 : 0)
+                      + (scene.radarState.getLeadsCenter()).size();
+            snprintf(val, sizeof(val), "%d", dat);
+            snprintf(unit, sizeof(unit), "cars");
+            break;}
+
+          case UIMeasure::TRAFFIC_COUNT_STOPPED:
+            {
+            snprintf(name, sizeof(name), "STOPPED");
+            int dat = scene.lead_vertices_stopped.size()
+                      + (scene.lead_data[0].getStatus() 
+                        && scene.lead_data[0].getVLeadK() < 3 
+                          ? 1 + (scene.radarState.getLeadsCenter()).size() 
+                          : 0);
+            snprintf(val, sizeof(val), "%d", dat);
+            snprintf(unit, sizeof(unit), "cars");
+            break;}
+
+          case UIMeasure::TRAFFIC_COUNT_ADJACENT_ONGOING:
+            {
+            snprintf(name, sizeof(name), "ADJ ONGOING");
+            int dat1 = scene.lateral_plan.getTrafficCountLeft();
+            int dat2 = scene.lateral_plan.getTrafficCountRight();
+            snprintf(val, sizeof(val), "%d:%d", dat1, dat2);
+            snprintf(unit, sizeof(unit), "cars");
+            break;}
+
+          case UIMeasure::TRAFFIC_ADJ_ONGOING_MIN_DISTANCE:
+            {
+            snprintf(name, sizeof(name), "MIN ADJ SEP");
+            float dat1 = scene.lateral_plan.getTrafficMinSeperationLeft();
+            float dat2 = scene.lateral_plan.getTrafficMinSeperationRight();
+            snprintf(val, sizeof(val), "%.1f:%.1f", dat1, dat2);
+            snprintf(unit, sizeof(unit), "s");
+            break;}
+
           case UIMeasure::LEAD_TTC:
             {
             snprintf(name, sizeof(name), "TTC");
@@ -1898,10 +2038,10 @@ static void ui_draw_measures(UIState *s){
               }
               else {
                 snprintf(val, sizeof(val), "%d", scene.engineRPM);
-                if (temp < 87){
+                if (temp < 74){
                   unit_color = nvgRGBA(84, 207, 249, 200); // cyan if too cool
                 }
-                else if (temp > 120){
+                else if (temp > 115){
                   unit_color = nvgRGBA(255, 0, 0, 200); // red if too hot
                 }
                 else if (temp > 105){
@@ -1921,10 +2061,10 @@ static void ui_draw_measures(UIState *s){
               }
               else {
                 snprintf(val, sizeof(val), "%d", scene.engineRPM);
-                if (temp < 190){
+                if (temp < 165){
                   unit_color = nvgRGBA(84, 207, 249, 200); // cyan if too cool
                 }
-                else if (temp > 250){
+                else if (temp > 240){
                   unit_color = nvgRGBA(255, 0, 0, 200); // red if too hot
                 }
                 else if (temp > 220){
@@ -1941,10 +2081,10 @@ static void ui_draw_measures(UIState *s){
               int temp = scene.car_state.getEngineCoolantTemp();
               snprintf(val, sizeof(val), "%d", temp);
               if(scene.engineRPM > 0 || temp >= 55) {
-                if (temp < 87){
+                if (temp < 74){
                   val_color = nvgRGBA(84, 207, 249, 200); // cyan if too cool
                 }
-                else if (temp > 120){
+                else if (temp > 115){
                   val_color = nvgRGBA(255, 0, 0, 200); // red if too hot
                 }
                 else if (temp > 105){
@@ -1961,10 +2101,10 @@ static void ui_draw_measures(UIState *s){
               int temp = int(float(scene.car_state.getEngineCoolantTemp()) * 1.8 + 32.5);
               snprintf(val, sizeof(val), "%d", temp);
               if(scene.engineRPM > 0 || temp >= 130) {
-                if (temp < 190){
+                if (temp < 165){
                   val_color = nvgRGBA(84, 207, 249, 200); // cyan if too cool
                 }
-                else if (temp > 250){
+                else if (temp > 240){
                   val_color = nvgRGBA(255, 0, 0, 200); // red if too hot
                 }
                 else if (temp > 220){
@@ -2437,11 +2577,25 @@ static void ui_draw_measures(UIState *s){
               snprintf(name, sizeof(name), "LANE W");
               if (s->is_metric){
                 snprintf(unit, sizeof(unit), "m");
-                snprintf(val, sizeof(val), "%f.1", scene.lateralPlan.laneWidth);
+                snprintf(val, sizeof(val), "%.1f", scene.lateralPlan.laneWidth);
               }
               else{
                 snprintf(unit, sizeof(unit), "ft");
                 snprintf(val, sizeof(val), "%.1f", scene.lateralPlan.laneWidth * 3.281);
+              }
+            }
+            break;
+
+          case UIMeasure::LANE_DIST_FROM_CENTER: 
+            {
+              snprintf(name, sizeof(name), "LANE CENTER");
+              if (s->is_metric){
+                snprintf(unit, sizeof(unit), "m");
+                snprintf(val, sizeof(val), "%.1f", scene.lateralPlan.laneCenter);
+              }
+              else{
+                snprintf(unit, sizeof(unit), "ft");
+                snprintf(val, sizeof(val), "%.1f", scene.lateralPlan.laneCenter * 3.281);
               }
             }
             break;
@@ -2490,7 +2644,7 @@ static void ui_draw_measures(UIState *s){
         
         int vallen = strlen(val);
         if (vallen > 4){
-          val_font_size -= (vallen - 4) * 5;
+          val_font_size -= (vallen - 4) * 8;
         }
         int unitlen = strlen(unit);
         if (unitlen > 5){
@@ -2640,7 +2794,8 @@ static void ui_draw_vision_event(UIState *s) {
   s->scene.wheel_touch_rect = {1,1,1,1};
   if (s->scene.engageable) {
     // draw steering wheel
-    const float rot_angle = -s->scene.angleSteers * 0.01745329252;
+    const float rot_angle_multiplier = s->scene.car_state.getVEgo() / 5.0;
+    const float rot_angle = -s->scene.angleSteers * 0.01745329252 * (rot_angle_multiplier > 1.0 ? rot_angle_multiplier : 1.0);
     const int radius = 88;
     const int center_x = s->fb_w - radius - bdr_s * 2;
     const int center_y = radius  + (bdr_s * 1.5);
@@ -2721,14 +2876,38 @@ static void ui_draw_vision_event(UIState *s) {
     }
   }
 
-  // current road name
-  if (s->scene.network_strength > 0 && !s->scene.map_open){//} && s->scene.current_road_name != ""){
-    nvgBeginPath(s->vg);
-    nvgTextAlign(s->vg, NVG_ALIGN_CENTER | NVG_ALIGN_TOP);
-    nvgFontFace(s->vg, "sans-regular");
-    nvgFontSize(s->vg, 75);
-    nvgFillColor(s->vg, COLOR_WHITE_ALPHA(255));
-    nvgText(s->vg, s->fb_w / 2, bdr_s - 31, s->scene.current_road_name.c_str(), NULL);
+  // current road name and heading
+
+  nvgBeginPath(s->vg);
+  nvgTextAlign(s->vg, NVG_ALIGN_CENTER | NVG_ALIGN_TOP);
+  nvgFontFace(s->vg, "sans-regular");
+  nvgFontSize(s->vg, 75);
+  nvgFillColor(s->vg, COLOR_WHITE_ALPHA(255));
+  char val[16];
+  if (s->scene.bearingAccuracy != 180.00) {
+    if (((s->scene.bearingDeg >= 337.5) && (s->scene.bearingDeg <= 360)) || ((s->scene.bearingDeg >= 0) && (s->scene.bearingDeg <= 22.5))) {
+      snprintf(val, sizeof(val), "(N)");
+    } else if ((s->scene.bearingDeg > 22.5) && (s->scene.bearingDeg < 67.5)) {
+      snprintf(val, sizeof(val), "(NE)");
+    } else if ((s->scene.bearingDeg >= 67.5) && (s->scene.bearingDeg <= 112.5)) {
+      snprintf(val, sizeof(val), "(E)");
+    } else if ((s->scene.bearingDeg > 112.5) && (s->scene.bearingDeg < 157.5)) {
+      snprintf(val, sizeof(val), "(SE)");
+    } else if ((s->scene.bearingDeg >= 157.5) && (s->scene.bearingDeg <= 202.5)) {
+      snprintf(val, sizeof(val), "(S)");
+    } else if ((s->scene.bearingDeg > 202.5) && (s->scene.bearingDeg < 247.5)) {
+      snprintf(val, sizeof(val), "(SW)");
+    } else if ((s->scene.bearingDeg >= 247.5) && (s->scene.bearingDeg <= 292.5)) {
+      snprintf(val, sizeof(val), "(W)");
+    } else if ((s->scene.bearingDeg > 292.5) && (s->scene.bearingDeg < 337.5)) {
+      snprintf(val, sizeof(val), "(NW)");
+    }
+  }
+  if (s->scene.network_strength > 0 && !s->scene.map_open){
+    nvgText(s->vg, s->fb_w / 2, bdr_s - 31, (s->scene.current_road_name + " " + val + " ").c_str(), NULL);
+  }
+  else {
+    nvgText(s->vg, s->fb_w / 2, bdr_s - 31, val, NULL);
   }
 }
 
@@ -2813,22 +2992,16 @@ static void ui_draw_vision_power_meter(UIState *s) {
         int h_pitch = h_rr + hu * pitch_power / s->scene.power_max[1];
         nvgBeginPath(s->vg);
         nvgRect(s->vg, xi+2, y_mid - h_rr - h_pitch / 2, wi-4, h_pitch);
-        nvgFillColor(s->vg, nvgRGBA(0, 140, 255, inner_fill_alpha/2));
-        nvgFill(s->vg);
         nvgStrokeWidth(s->vg, 5);
         nvgStrokeColor(s->vg, COLOR_WHITE_ALPHA(150));
         nvgStroke(s->vg);
         nvgBeginPath(s->vg);
         nvgRect(s->vg, xi+2, y_mid - h_drag - h_rr / 2, wi-4, h_rr);
-        nvgFillColor(s->vg, nvgRGBA(0, 180, 255, inner_fill_alpha/2));
-        nvgFill(s->vg);
         nvgStrokeWidth(s->vg, 5);
         nvgStrokeColor(s->vg, COLOR_WHITE_ALPHA(150));
         nvgStroke(s->vg);
         nvgBeginPath(s->vg);
         nvgRect(s->vg, xi+2, y_mid - h_drag / 2, wi-4, h_drag);
-        nvgFillColor(s->vg, nvgRGBA(0, 230, 255, inner_fill_alpha/2));
-        nvgFill(s->vg);    
         nvgStrokeWidth(s->vg, 5);
         nvgStrokeColor(s->vg, COLOR_WHITE_ALPHA(150));
         nvgStroke(s->vg);
@@ -2864,22 +3037,16 @@ static void ui_draw_vision_power_meter(UIState *s) {
         int h_pitch = h_rr + hu * pitch_power / s->scene.power_max[0];
         nvgBeginPath(s->vg);
         nvgRect(s->vg, xi+2, y_mid - h_rr - h_pitch / 2, wi-4, h_pitch);
-        nvgFillColor(s->vg, nvgRGBA(249,240,100,inner_fill_alpha/2));
-        nvgFill(s->vg);
         nvgStrokeWidth(s->vg, 5);
         nvgStrokeColor(s->vg, COLOR_WHITE_ALPHA(150));
         nvgStroke(s->vg);
         nvgBeginPath(s->vg);
         nvgRect(s->vg, xi+2, y_mid - h_drag - h_rr / 2, wi-4, h_rr);
-        nvgFillColor(s->vg, nvgRGBA(249,240,150,inner_fill_alpha/2));
-        nvgFill(s->vg);
         nvgStrokeWidth(s->vg, 5);
         nvgStrokeColor(s->vg, COLOR_WHITE_ALPHA(150));
         nvgStroke(s->vg);
         nvgBeginPath(s->vg);
         nvgRect(s->vg, xi+2, y_mid - h_drag / 2, wi-4, h_drag);
-        nvgFillColor(s->vg, nvgRGBA(249,240,200,inner_fill_alpha/2));
-        nvgFill(s->vg);    
         nvgStrokeWidth(s->vg, 5);
         nvgStrokeColor(s->vg, COLOR_WHITE_ALPHA(150));
         nvgStroke(s->vg);