+++ selfdrive/ui/ui.h
@@ -39,8 +39,8 @@
 #define COLOR_GREEN_ALPHA(x) nvgRGBA(0, 255, 0, x)
 #define COLOR_BLUE nvgRGBA(0, 0, 255, 255)
 #define COLOR_BLUE_ALPHA(x) nvgRGBA(0, 0, 255, x)
-#define COLOR_GRACE_BLUE nvgRGBA(0,4,167, 255)
-#define COLOR_GRACE_BLUE_ALPHA(x) nvgRGBA(0,4,167, x)
+#define COLOR_GRACE_BLUE nvgRGBA(0,100,255, 255)
+#define COLOR_GRACE_BLUE_ALPHA(x) nvgRGBA(0,100,255, x)
 #define COLOR_ORANGE nvgRGBA(255, 175, 3, 255)
 #define COLOR_ORANGE_ALPHA(x) nvgRGBA(255, 175, 3, x)
 #define COLOR_YELLOW_ALPHA(x) nvgRGBA(218, 202, 37, x)
@@ -142,7 +142,7 @@ const QColor bg_colors [] = {
 const QColor alt_bg_colors [] = {
   [STATUS_DISENGAGED] =  QColor(0x00, 0x30, 0x4A, 0xc8),
   [STATUS_ENGAGED] = QColor(0x00, 0x04, 0xA7, 0xf1),
-  [STATUS_WARNING] = QColor(0x6A, 0x00, 0x7E, 0xf1),
+  [STATUS_WARNING] = QColor(0xAB, 0x00, 0xAE, 0xf1),
   [STATUS_ALERT] = QColor(0x92, 0x00, 0x0D, 0xf1),
 };
 
@@ -207,8 +207,12 @@ typedef enum UIMeasure { //rearrange here to adjust order when cycling measures
   ROLL,
   ROLL_DEVICE,
   LANE_WIDTH,
+  LANE_DIST_FROM_CENTER,
   DISTANCE_TRAVELLED,
-  // Lead info
+  INTERACTION_TIMER,
+  INTERVENTION_TIMER,
+  DISTRACTION_TIMER,
+  // Lead/traffic info
   FOLLOW_LEVEL,
   LEAD_TTC,
   LEAD_DISTANCE_LENGTH,
@@ -218,6 +222,14 @@ typedef enum UIMeasure { //rearrange here to adjust order when cycling measures
   LEAD_COSTS,
   LEAD_VELOCITY_RELATIVE,
   LEAD_VELOCITY_ABS,
+  LANE_POSITION,
+  LANE_OFFSET,
+  TRAFFIC_COUNT_TOTAL,
+  TRAFFIC_COUNT_ONCOMING,
+  TRAFFIC_COUNT_ONGOING,
+  TRAFFIC_COUNT_STOPPED,
+  TRAFFIC_COUNT_ADJACENT_ONGOING,
+  TRAFFIC_ADJ_ONGOING_MIN_DISTANCE,
   // EV info
   HVB_VOLTAGE,
   HVB_CURRENT,
@@ -480,6 +492,7 @@ typedef struct UIScene {
   struct _LateralPlan
   {
     float laneWidth;
+    float laneCenter;
 
     float dProb;
     float lProb;